its work

+++ client/src/pages/Admin/UpdateProduct.js
@@ -87,6 +87,7 @@ const UpdateProduct = () => {
     const handleDelete = async() =>{
        try {
         let answer = window.prompt('Are You Sure want to delete this product ?')
+        if(!answer) return;
         const {data} = await axios.delete(`/api/v1/product/delete-product/${id}`);
         toast.success('Product Deleted Successfully');
         navigate('/dashboard/admin/products')
